Safely destroy channels on disconnect
Previously on disconnect we'd quit all channels which would automatically
destroy the channel if it was now empty. This caused an InvalidOperationException
as we continued to iterate over the channels list.

Fixed by creating a list of channels to later iterate over and destroy if
appropriate.

Note that this bug was causing users to not be disconnected cleanly; they'd
remain referenced to in SocketListener, likely until they timed out
(may have even survived that).