IRCD-13 - Simplify iterating over all clients
Each ServerListener contains a list of its clients; we add a reference to this
to IRCd.ClientLists. We modify this list as appropriate.

This meant that iterating through clients looked something like:
	foreach (var clientList in IRCd.ClientLists) {
		foreach (var client in clientList) {
			// ...
		}
	}

This is unnecessarily verbose (required because in this case we want
to iterate through every single client on the server, so every single
listener's clients).

With this change, it becomes:
	foreach (var client in IRCd.Clients) {
		// ...
	}

This is far clearer and makes the intent a bit more visible too.

h/t lewis for the .SelectMany() suggestion. This has been bothering
me for some time.