IRCD-13 - Tolerate absent colon in CAP REQ
Messages should be of the form:
        CAP REQ :<cap1> <cap2> ...

However it is not hard to support
        CAP REQ <cap1> <cap2> ...
even if this is not mentioned in the spec.

One of the examples in the spec has
        CAP REQ <cap1>
so we choose to tolerate that.

We also send a NOTICE letting the user know
to report this as a bug to the client
developer.